requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioice==0.9.0
aioresponses==0.7.6
aiortc==1.9.0
aiosignal==1.3.1
altair==4.2.2
altgraph==0.17.4
annotated-types==0.6.0
anthropic==0.34.2
anyio==3.7.1
appdirs==1.4.4
APScheduler==3.10.1
asgiref==3.7.2
async-property==0.2.2
async-timeout==4.0.3
asyncio-dgram==2.2.0
asyncstdlib==3.10.9
attrs==23.2.0
Authlib==1.4.1
av==12.3.0
awscli==1.33.44
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.123
botocore==1.34.162
Brotli==1.1.0
browser-cookie3==0.19.1
bs4==0.0.2
cachelib==0.13.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
colorama==0.4.4
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==43.0.3
cssselect==1.2.0
curl-cffi==0.5.10
cycler==0.12.1
Cython==3.0.0
dataclasses-json==0.6.7
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker==6.1.3
docutils==0.16
duckduckgo-search==4.1.0
entrypoints==0.4
exceptiongroup==1.2.0
fake-useragent==1.4.0
fastapi==0.105.0
filelock==3.16.1
Flask==3.0.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-Mail==0.10.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.0
fsspec==2024.10.0
g4f==0.1.9.3
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gpt4all==2.0.2
GPUtil==1.4.0
greenlet==3.1.1
grpcio==1.60.0
gspread==5.12.3
h11==0.14.0
hatch==1.13.0
hatchling==1.25.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
huggingface-hub==0.26.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
ifaddr==0.2.0
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
inference==0.26.0
iniconfig==2.0.0
isodate==0.7.2
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.2
jiter==0.7.0
jmespath==1.0.1
joblib==1.4.2
Js2Py==0.74
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keyring==25.5.0
kiwisolver==1.4.7
lazy_loader==0.4
llm==0.11.1
loguru==0.7.2
lxml==5.1.0
lz4==4.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.23.1
matplotlib==3.9.2
mcstatus==11.1.1
mdurl==0.1.2
mmh3==4.0.1
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
mutagen==1.47.0
mypy-extensions==1.0.0
names==0.3.0
nest-asyncio==1.5.8
networkx==3.4.2
nose2==0.14.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==2.1.0
onnxruntime==1.19.2
openai==1.54.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parse==1.20.1
pathspec==0.12.1
pefile==2023.2.7
pendulum==3.0.0
pexpect==4.9.0
piexif==1.1.3
pillow==10.2.0
platformdirs==4.1.0
playwright==1.48.0
pluggy==1.5.0
prettytable==3.12.0
proglog==0.1.10
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.21.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.2
psutil==6.1.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
PuLP==2.9.0
py-arkose-generator==0.0.0.2
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybase64==1.4.0
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.8.1b0
pydot==3.0.2
pydub==0.25.1
pyee==12.0.0
PyExecJS==1.5.1
pygame==2.6.1
Pygments==2.17.2
pygpt4all==1.1.0
pygptj==2.0.3
pyht==0.0.18
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pyjsparser==2.7.1
pylibsrtp==0.10.0
pyllamacpp==2.4.2
Pympler==1.0.1
pyOpenSSL==24.2.1
pyparsing==3.1.1
PyPasser==0.0.5
pyperclip==1.9.0
pyppeteer==1.0.2
pyquery==2.0.0
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.3.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==20.7
python-ulid==2.2.0
pytube==15.0.0
pytz==2023.3.post1
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.1
redis==5.2.0
referencing==0.32.0
requests==2.31.0
requests-auth-aws-sigv4==0.7
requests-file==2.1.0
requests-html==0.10.0
requests-oauthlib==1.1.0
requests-toolbelt==1.0.0
rich==13.5.2
rpds-py==0.16.2
rsa==4.7.2
s3transfer==0.10.3
schedule==1.2.2
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.17.2
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
skypilot==0.5.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.35.2
starlette==0.27.0
streamlit==1.21.0
structlog==24.4.0
supervision==0.22.0
sympy==1.13.3
tabulate==0.9.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktok-uploader==1.0.16
tiktokautouploader==3.7
time-machine==2.16.0
tldextract==5.1.3
tls-client==0.2.2
tokenizers==0.20.3
toml==0.10.2
tomli_w==1.1.0
tomlkit==0.13.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
trove-classifiers==2024.10.21.16
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
undetected-chromedriver @ git+https://github.com/ultrafunkamsterdam/undetected-chromedriver.git@0aa5fbe252370b4cb2b95526add445392cad27ba
uritemplate==4.1.1
urllib3==1.26.20
userpath==1.9.2
uv==0.5.0
uvicorn==0.24.0.post1
validators==0.22.0
virtualenv==20.27.1
w3lib==2.2.1
waitress==2.1.2
watchdog==3.0.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.7.0
websockets==10.4
Werkzeug==3.0.1
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.17.1
youtube-transcript-api==0.6.2
yt-dlp==2024.10.22
zipp==3.17.0
zstandard==0.23.0
zxing-cpp==2.2.0
